refactor: address quick cast review feedback
